+++ controllers/categoriesController.js
@@ -35,31 +35,6 @@ const categoriesController = {
     const items = await getItems(req.params);
     res.render("category", { title: category, category, categories, items });
   }),
-  getCategoryEdit: asyncHandler(async (req, res) => {}),
-  getCategoryDelete: asyncHandler(async (req, res) => {
-    console.log("getCategoryDelete running...");
-    console.log("req.url:", req.url);
-    console.log("req.params:", req.params);
-    const categories = await getCategories();
-    const { category } = req.params;
-    res.render("delete", { title: "Delete Category", categories });
-  }),
-  postCategoryDelete: asyncHandler(async (req, res) => {
-    console.log("postCategoryDelete running...");
-    console.log("req.url:", req.url);
-    console.log("req.query:", req.query);
-    console.log("req.params:", req.params);
-    console.log("req.body:", req.body);
-    // Run db query to delete category
-    // Which makes more sense, update items before deleting category or
-    // deleting category before updating items?
-    // How to show a confirmation modal?
-    // Does it make sense to define a modal on the server side?
-    // await updateItems(req.params);
-    // await deleteCategory(req.params);
-    // Redirect vs render
-    res.redirect("/categories");
-  }),
 };
 
 module.exports = categoriesController;